Update day03
Parts 1 and 2.

Surprisingly easy to convert to part 2. Still brute forces through string, just added another line to look out for "do()" and "don't()" instructions, and moved the main code to a method so that it could be called with an int flag to determine if it should look for the "do()" or "don't()" instructions. Also renamed some variables to be more in line with the story.