Create day04
Had a bit of a tumultuous week with some unexpected things so was unable to submit for a few days. Then will be going on a vacation so will look pretty bare.

A bit of a mess, and only Part 1. Initially had 4 different methods, but realized that they were largely repetitive and could be combined, so combined them into main, but the horizontal check method is still there. Not the most elegant solution, but the side-idea I started with the diagonal will be the basis of the Part 2.