Update day04
Added Part 2, was actually simple comapred to Part 1, thought I'd have to do some kind of recursive searching, but it's just a different set of 4 chars to check. Could rewrite it to only go through it one time as opposed to every time, but would need constant modifiers based on the positioning. 

Absorbed the horizontal checking into the main method, so it's all in one now.