Create day06
Skipped to day 6, still recovering from jetlag and a few other things.

Could calculate the obstructions directly from the file and save some of the memory, but I like to copy the file then close it to make sure the data is intact. Variation of a few grid path-ing problems from years past.