Update day06
Now updated with Part 1 and Part 2. 

Really overthought this one, was going to put obstacles while building the visit list for Part 1 while then running a recursive method to see if that looped, but then realized that each obstacle would only need to be on the initial path otherwise the guard would never reach it, and it's easier to use the first Part as a list for the potential obstacle spots.

Moved some things to methods, but there's a bit of redundancy, both pairs do roughly the same thing, but with slightly different considerations and outputs.

A better method would likely be to use the obstacles to figure it out as opposed to "running" every single one, you'd just want to check if there are 3 other pre-existing obstacles that create a box.