Create day07
Took a while to figure out the recursion on this one, realized I was going the wrong way trying to return long; confused me with needing to run 2 threads but return 1 thing. Boolean fixed it. Other minor hiccup was making sure your regex works with ": " and doesn't create an empty string.