Create day08
Wasn't feeling well yesterday so took a break. Bit of a throwback trying to remember my geometry days, and making sure to keep track of the direction the line was facing. Realized it was unnecessary to take the absolute value and after printing out some smaller test inputs was able to narrow down the proper equations to generate the antinodes.