Create day09 - legacy
This is what I get for not looking at the sample data before diving in.

This code works for the example given and does a bruteforce generation + swap of the digits. It fails because it only works on single digit file IDs, which is not true for the actual data set. Could be redone to work for multi digit IDs, but is already cumbersome at a smaller input size.