Update day09
Part 1 and Part 2, after a week long vacation.

Part 1 builds "blocks" arrayList for part 2, and part 2 reuses some variables that are reset on lines 60-63. I also added explicit casts to long for the checkSum calculations to increase legibility.

The part 2 code takes much longer since it builds the entire file system. I initially tried a 2D array, which failed pretty quickly when I realized that partial blocks could form and I would need to increase the size. This new approach uses an arrayList to allow for dynamic sizing. It's dangerous in that indexing can be broken, but it works out ok.

There is a boolean flag system that updates the first empty space index to prevent the program from always searching from the beginning. At worst case there is always a small space in the beginning that never gets filled, which would add some extra time as it checks. However, in my given data set, it ended up cutting the runtime for p2 in almost half.

I changed the runtime calculations to get a more exact runtime for p2. P1 takes on average > 10ms, so it's not really affecting p2 much, but for future consistency I'll start using this framework.