Create day10
Part 1.

Fairly simple recursive pathing algorithm, trying to change to more modern conventions regarding declaring variables as late as possible as opposed to all at the top. Learned it was not a good idea to use arrays as keys, given the nightmare between matching the array vs the object, so it's all converted to strings.