Update day11
Part 1 and Part 2.

Completely rewritten, as expected the runtime grew exponentially with each blink. Since the position of each stone is actually irrelevant, what I've done is keep track of the number of repeat stones instead of keeping track of individual stones. Since there are a lot repeating patterns it ends with a much much more manageable runtime. I kept the legacy code underneath; the new blink method uses largely the same code. I've rewritten it to be more readable and less redundant even though it's less compact.

It blinks 25 times for Part 1, then the remaining 50 for Part 2, so I've combined the runtime calculation. Also realized that during blinks I was parsing the split String to Long, then back to String, without needing them in Long at any point, so trimmed that out.