Update day12
Part 1 Complete.

As with any good recursion, spent all day debugging; chasing down variables and  exit conditions. Found that my code was not checking all sides due to the way it recursed and exited, so rewrote that. This is probably a prime candidate for a custom object, as opposed to use a bunch of hashmaps and sets, but it works. Not sure how it will work for part 2 though.