Create day13
Skipping ahead for now.

The initial file read does a lot of heavy lifting, where it splits each block into an array. The splitting is pretty ugly with a substring, and a regex. It should probably be split up more for readability.

Hit a bit of a wall with it, there are a lot of implied rules that are not explicitly stated, so it was over-processing a lot of the numbers thinking there would potentially be a lot of edge cases. Read from other people on the r/AdventOfCode subreddit that it was not the case which simplified it a lot. 

I was also only checking to see if bX and bY could be used to match the prize after aX and aY, but not that they would be the same amount of presses. This obviously was a big oversight and gave me lots of false positives and too low token values. It took me a while to realize, which was most of the actual time I spent trying to figure this out.