Create day14
Part 1.

Went through a few options, initially looped and built a list of all visited points before returning to the start, then looked for the 100th position on that list (with modulo). This ended up working well for the example. but I noticed that for the data that each one was looping far beyond 100 seconds after capping the loop to 100 iterations . Changed to just do it with math instead.

Then spent an embarrassing amount of time having problems because I incorrectly indexed the "middle" lines. Thought I was doing something wrong with the modulo or the indexing for "seconds" vs the loop. Constantly switching back and forth from the example to the test data didn't help either.