Update day14
Unfinished Part 2.

Working on a solution that uses the initial setup I had with bruteforcing a list before diving into it to build boards.

Was attempting to verify it visually, but looking at 10k+ boards gets old fast. Attempting to use a hybrid approach that both builds boards and writes to a file, while calculating some kind of zone based number analysis. Need to code the number properly, standard deviation or something similar.

Will probably keep the hybrid method, since verifying with a visual component is important, but can probably just loop through it with the math based Part 1 method instead of generating brute force lists. This also assumes the lists are properly combined, will have to use Part 1 to verify when I return to it.