Update day14
Parts 1 and 2. Runs at the same time.

Rewritten again. The same thought process as the first part; brute force into math based. Originally brute forced lists and then ran through them, realized after too long that I was still doing the %100 from the part 1 code which was causing errors. Decided to revamp it to use the same part 1 math based version. 

The only problem with that is that you need to pick a sufficiently long loop or be able to follow each loop and cancel it when it's found. I've used the maximum bot loop, which you could only know after brute forcing it. You could track a single bot till it loops, but you'd have to assume you correctly picked a bot with a long enough loop. While most of them take the same time, there were a few that seemed to loop much much faster (this could also be an error based on the original methods problems).

After rewriting it, I accidentally stumbled on the correct solution. Somehow, by using 9 sections but only defining the board by 6-having 3 extra "0" sections-and also taking the lowest standard deviation, it produced the correct answer. Fixed it to produce the answer with the proper 6 sections and highest SD. You then go verify the data by looking at the file to see the "picture". I was using SD after looking at the AOC subreddit, but the lowest safety factor works as well. I've left the safety factor approach in the code since it's the easiest to follow and matches the original prompt.

This hybrid approach of coding + visual confirmation was interesting, but I found it initially frustrating. Part 1 does suggest the solution to you by having you look at quadrants and take the "safety factor", but doesn't really provide much else or even tell you what you're looking for. This might be easy enough for more proficient coders or people using a visual language, but was confusing for me.