Update day15
Part 1.

Rewrote it twice. Corrected the mistake from indexing printouts (now removed) using the full file size forgetting that there was the path line under the map. Initially thought that the robot would move until it hit a wall or the boxes it was pushing hit a wall, which was clearly not the case. Then realized that it only needs to attempt to move one space, which was actually easier to write.