Create day19 - partial
Jumped to day 19. Day 17p2 requires some non programming thinking, which I didn't want to get into on a friday. Day 18 was another grid, so here we are at 19.

My thinking is to remove "extra" stripe blocks that are made up of smaller towel blocks that exist first. I feel like this will bite me in p2 if towels become limited.Either way, it doesn't work. It's producing the correct result for the example, which is much much smaller than the actual data. Has a lot of prints to help debug what's happening. Need to create a medium sized test case to make sure it works properly.