Update and rename day21 to day22
Part 1 and Part 2.

Mistakenly called it Day 21, where it's actually Day 22 puzzle. Fixed.

Part 2 was not super hard to implement, though reading comprehension is needed. Over-complicated it again by thinking I needed the best price if the sequence repeated, but I just needed the first price.

Abstracted the key string a bit to keep it fixed length so that I could use a StringBuilder instead of creating a string from a list every iteration, which cut the time nearly in half. You don't need the "winning" sequence in the end so no need to preserve it.