Create day23
Part 1.

Not sure if I prefer building the adjacency list with the file open, or copying to a string list then work on that. This does save some time though. Otherwise fairly easy brute force. Use list for edges for easier indexing, but then use set to keep track because of the inherent ability to filter out duplicates.